chore(transcription): move benchmark out of tests
chore(transcription): remove transcription workflow

docs(transcription): add benchmark info
