--- packages/tests/src/transcription/benchmark.spec.ts
@@ -1,139 +1,0 @@
-import { createLogger } from 'winston'
-import { join } from 'node:path'
-import { performance, PerformanceObserver } from 'node:perf_hooks'
-import { tmpdir } from 'node:os'
-import { rm, mkdir } from 'node:fs/promises'
-import { buildAbsoluteFixturePath, buildSUUID, SUUID } from '@peertube/peertube-node-utils'
-import {
-  transcriberFactory,
-  TranscriptFile,
-  TranscriptFileEvaluator,
-  TranscriptionEngine,
-  TranscriptionModel
-} from '@peertube/peertube-transcription'
-import { millisecondsToTime } from '@peertube/peertube-core-utils'
-
-interface TestResult {
-  uuid: SUUID
-  WER?: number
-  CER?: number
-  duration?: number
-  engine?: TranscriptionEngine
-  model?: string
-}
-
-type Benchmark = Record<SUUID, TestResult>
-
-const benchmarkReducer = (benchmark: Benchmark = {}, testResult: TestResult) => ({
-  ...benchmark,
-  [testResult.uuid]:  {
-    ...benchmark[testResult.uuid],
-    ...testResult
-  }
-})
-
-interface FormattedTestResult {
-  WER?: string
-  CER?: string
-  duration?: string
-  model?: string
-  engine?: string
-}
-
-const formatTestResult = ({ WER, CER, duration, engine, model }: Partial<TestResult>): FormattedTestResult => ({
-  WER: WER ? `${WER * 100}%` : undefined,
-  CER: CER ? `${CER * 100}%` : undefined,
-  duration: duration ? millisecondsToTime(duration) : undefined,
-  model,
-  engine: engine.name
-})
-
-describe('Transcribers benchmark', function () {
-  const transcribers = [
-    'openai-whisper',
-    'whisper-ctranslate2',
-    'whisper-timestamped'
-  ]
-  const models = [
-    'tiny',
-    'small'
-  ]
-
-  const transcriptDirectory = join(tmpdir(), 'peertube-transcription/benchmark/')
-  const mediaFilePath = buildAbsoluteFixturePath('transcription/videos/derive_sectaire.mp4')
-  const referenceTranscriptFile = new TranscriptFile({
-    path: buildAbsoluteFixturePath('transcription/videos/derive_sectaire.txt'),
-    language: 'fr',
-    format: 'txt'
-  })
-
-  let benchmark: Record<string, TestResult> = {}
-
-  before(async function () {
-    await mkdir(transcriptDirectory, { recursive: true })
-
-    const performanceObserver = new PerformanceObserver((items) => {
-      items
-        .getEntries()
-        .forEach((entry) => {
-          benchmark = benchmarkReducer(benchmark, {
-            uuid: entry.name as SUUID,
-            duration: entry.duration
-          })
-        })
-    })
-    performanceObserver.observe({ type: 'measure' })
-  })
-
-  transcribers.forEach(function (transcriberName) {
-    describe(`Create ${transcriberName} transcriber for the benchmark`, function () {
-      const transcriber = transcriberFactory.createFromEngineName(
-        transcriberName,
-        createLogger(),
-        transcriptDirectory
-      )
-
-      models.forEach((modelName) => {
-        it(`Run ${transcriberName} transcriber benchmark with ${modelName} model`, async function () {
-          this.timeout(15 * 1000 * 60) // 15 minutes
-          const model = new TranscriptionModel(modelName)
-          const uuid = buildSUUID()
-          const transcriptFile = await transcriber.transcribe({
-            mediaFilePath,
-            model,
-            language: 'fr',
-            format: 'txt',
-            runId: uuid
-          })
-          const evaluator = new TranscriptFileEvaluator(referenceTranscriptFile, transcriptFile)
-          await new Promise(resolve => setTimeout(resolve, 1))
-
-          benchmark = benchmarkReducer(benchmark, {
-            uuid,
-            engine: transcriber.engine,
-            WER: await evaluator.wer(),
-            CER: await evaluator.cer(),
-            model: model.name
-          })
-        })
-      })
-    })
-  })
-
-  after(async function () {
-    const benchmarksGroupedByModel = Object
-        .keys(benchmark)
-        .reduce((benchmarksGroupedByModel, uuid, currentIndex, array) => ({
-          ...benchmarksGroupedByModel,
-          [benchmark[uuid].model]: {
-            ...benchmarksGroupedByModel[benchmark[uuid].model],
-            [uuid]: formatTestResult(benchmark[uuid])
-          }
-        }), {})
-    Object.values(benchmarksGroupedByModel).forEach(benchmark => console.table(benchmark))
-
-    await rm(transcriptDirectory, { recursive: true, force: true })
-
-    performance.clearMarks()
-  })
-})
